BZSL-Python/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: Windows 10
python=3.7.10
numpy=1.19.2 
scipy=1.6.1
sklearn=1.0.2
tensorflow=2.1.0
biopython=1.79